+++ crates/shine-game/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod ai;
 pub mod application;
 pub mod camera_rig;
 pub mod input_manager;
+pub mod math;
