+++ crates/shine-game/src/math/jackknife_classifier/jackknife_error.rs
