+++ crates/shine-game/src/math/jackknife_classifier/jackknife_math.rs
@@ -43,59 +43,10 @@ pub trait JackknifePoint: Clone + ops::Index<usize, Output = f32> + ops::IndexMu
     fn lerp(a: &Self, b: &Self, t: f32) -> Self;
 }
 
-pub struct CostMatrix {
-    size: (usize, usize),
-    data: Vec<f32>,
-}
-
-impl CostMatrix {
-    pub fn new() -> Self {
-        CostMatrix { size: (0, 0), data: Vec::new() }
-    }
-
-    pub fn clear(&mut self) {
-        self.size = (0, 0);
-        self.data.clear();
-    }
-
-    pub fn init(&mut self, width: usize, height: usize) {
-        self.size = (width, height);
-        self.data.clear();
-        self.data.resize(height * width, f32::INFINITY);
-        self[(0, 0)] = 0.0;
-    }
-
-    pub fn size(&self) -> (usize, usize) {
-        self.size
-    }
-}
-
-impl ops::Index<(usize, usize)> for CostMatrix {
-    type Output = f32;
-
-    fn index(&self, index: (usize, usize)) -> &Self::Output {
-        let (i, j) = index;
-        debug_assert!(i < self.size.0 && j < self.size.1);
-        &self.data[i * self.size.1 + j]
-    }
-}
-
-impl ops::IndexMut<(usize, usize)> for CostMatrix {
-    fn index_mut(&mut self, index: (usize, usize)) -> &mut Self::Output {
-        let (i, j) = index;
-        debug_assert!(i < self.size.0 && j < self.size.1);
-        &mut self.data[i * self.size.1 + j]
-    }
-}
-
 pub trait JackknifePointMath<V>: JackknifePoint {
     fn path_length(points: &[V]) -> f32;
     fn resample(points: &[V], count: usize) -> Vec<V>;
     fn z_normalize(points: &mut [V]);
-
-    fn dtw<F>(cost: &mut CostMatrix, v1: &[V], v2: &[V], radius: usize, distance: F) -> f32
-    where
-        F: Fn(&V, &V) -> f32;
 }
 
 impl<V> JackknifePointMath<V> for V
@@ -193,29 +144,5 @@ where
         for point in points.iter_mut() {
             *point = V::from_sub(point, &mean).div_component(&var);
         }
-    }
-
-    fn dtw<F>(cost: &mut CostMatrix, v1: &[V], v2: &[V], radius: usize, distance: F) -> f32
-    where
-        F: Fn(&V, &V) -> f32,
-    {
-        let n = v1.len() + 1;
-        let m = v2.len() + 1;
-
-        cost.init(n, m);
-
-        for i in 1..n {
-            let range_min = if i > radius { i - radius } else { 1 };
-            let range_max = (i + radius).min(m - 1);
-            for j in range_min..=range_max {
-                let minimum: f32 = (cost[(i - 1, j)]) // repeat v1 element
-                    .min(cost[(i, j - 1)]) // repeat v2 element
-                    .min(cost[(i - 1, j - 1)]); // new element
-                cost[(i, j)] = minimum;
-                cost[(i, j)] += distance(&v1[i - 1], &v2[j - 1]);
-            }
-        }
-
-        cost[(n - 1, m - 1)]
     }
 }
