+++ crates/shine-game/src/math/mod.rs
