+++ crates/shine-game/src/math/unistroke_templates.rs
